🔧 (install scripts): add user confirmation prompt for Elixir LSP and UV installation ♻️ (install scripts): remove color formatting for cleaner script output ✨ (setup.sh): enable UV installation script by uncommenting its execution

+++ setup/scripts/install_elixir_lsp.sh
@@ -1,19 +1,11 @@
-#!/bin/bash
+if ask "Do you want to install Elixir LSP?"; then
+  LATEST_URL=$(curl -s https://api.github.com/repos/elixir-lsp/elixir-ls/releases/latest \
+    | jq -r '.assets[] | select(.name | test("elixir-ls.*\\.zip$")).browser_download_url')
+  curl -L -o elixir-ls.zip $LATEST_URL
 
-BOLD='\033[1m'
-GREEN='\033[32m'
-RESET='\033[0m'
-
-LATEST_URL=$(curl -s https://api.github.com/repos/elixir-lsp/elixir-ls/releases/latest \
-  | jq -r '.assets[] | select(.name | test("elixir-ls.*\\.zip$")).browser_download_url')
-curl -L -o elixir-ls.zip $LATEST_URL
-
-rm -rf /home/ton618/.local/share/elixir-lsp/elixir-ls
-
-mkdir -p /home/ton618/.local/share/elixir-lsp
-
-unzip elixir-ls.zip -d /home/ton618/.local/share/elixir-lsp/elixir-ls
-
-rm elixir-ls.zip
-
-echo -e "${BOLD}${GREEN}Elixir LSP is updated.${RESET}"
+  sudo rm -rf /home/ton618/.local/share/elixir-lsp/elixir-ls
+  sudo mkdir -p /home/ton618/.local/share/elixir-lsp
+  sudo unzip elixir-ls.zip -d /home/ton618/.local/share/elixir-lsp/elixir-ls
+  rm elixir-ls.zip
+  success "Elixir LSP installed!"
+fi
